Debugging Request Body Data

+++ backend/models/PassedModel.js
@@ -7,7 +7,7 @@ const Passed = dbase.define('rft_main',{
     line: DataTypes.STRING,
     status: DataTypes.STRING,
     code: DataTypes.STRING,
-    qty: DataTypes.SMALLINT,
+    qty: DataTypes.INTEGER,
 },{
     freezeTableName:true
 });
